# functions/docs/swagger.yaml
openapi: "3.0.0"
info:
  title: "전기안전 - Firebase Functions API"
  version: "1.0.0"
  description: |
    ## 🚀 Endpoint
    배포가 완료되면 아래 주소에서 API를 확인할 수 있습니다.

    - `https://us-central1-electrical-safety-4c9bd.cloudfunctions.net/api`
      <br>  <br>
    - `https://api-2hj4xnf47q-uc.a.run.app`
paths:
...
  /complex:
    post:
      tags:
        - Complex
      summary: "단지 추가 또는 아파트 추가"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - flag
              properties:
                flag:
                  type: string
                  description: "CPX: 단지 추가, APT: 아파트 추가"
                  enum: [CPX, APT]
                  example: "CPX"
                complexName:
                  type: string
                  description: |
                    "단지 추가 시 필수 입력입니다. 단지 이름을 입력하세요.  
                    예: '1단지'"
                  example: "1단지"
                apartments:
                  type: array
                  description: |
                    "아파트 추가 시 필수 입력입니다. 추가할 아파트 목록을 입력하세요.  
                    예: ['A 아파트', 'B 아파트']"
                  items:
                    type: string
                  example: ["A 아파트", "B 아파트"]
      responses:
        "201":
          description: "단지가 성공적으로 추가됨"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "단지가 성공적으로 추가되었습니다."
                  id:
                    type: string
                    example: "complexId123"
        "200":
          description: "아파트 목록이 성공적으로 업데이트됨"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "아파트 목록이 성공적으로 업데이트되었습니다."
                  id:
                    type: string
                    example: "complexId123"
        "400":
          description: "잘못된 요청"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "단지 이름은 필수 입력값입니다."
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "서버 오류 메시지"

    get:
      tags:
        - Complex
      summary: "모든 단지 목록 조회"
      responses:
        "200":
          description: "단지 목록 조회 성공"
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  properties:
                    name:
                      type: string
                      description: "단지 이름"
                      example: "1단지"
                    apartments:
                      type: array
                      items:
                        type: string
                      description: "단지에 포함된 아파트 목록"
                      example: ["A 아파트", "B 아파트"]
        "404":
          description: "등록된 단지가 없음"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "등록된 단지가 없습니다."
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "서버 오류 메시지"
  /complex/{id}:
    put:
      tags:
        - Complex
      summary: "단지 또는 아파트 수정"
      parameters:
        - name: id
          in: path
          required: true
          description: "수정할 단지의 ID"
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - flag
              properties:
                flag:
                  type: string
                  enum: [CPX, APT]
                  description: "CPX: 단지 수정, APT: 아파트 수정"
                  example: "CPX"
                complexName:
                  type: string
                  description: |
                    "단지 수정 시 사용할 새로운 이름을 입력하세요.  
                    예: '2단지'"
                  example: "2단지"
                apartments:
                  type: array
                  description: |
                    "단지에 포함된 아파트 목록을 수정하세요.  
                    예: ['C 아파트', 'D 아파트']"
                  items:
                    type: string
                  example: ["C 아파트", "D 아파트"]
                apartmentIndex:
                  type: integer
                  description: |
                    "아파트를 수정할 경우 인덱스를 입력하세요 (0부터 시작).  
                    예: 1"
                  example: 1
                apartmentName:
                  type: string
                  description: |
                    "수정할 아파트의 새 이름을 입력하세요.  
                    예: 'B 아파트'"
                  example: "B 아파트"
      responses:
        "200":
          description: "수정 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "단지/아파트 수정 성공"
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "서버 오류 메시지"

    delete:
      tags:
        - Complex
      summary: "단지 또는 아파트 삭제"
      parameters:
        - name: id
          in: path
          required: true
          description: "삭제할 단지의 ID"
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - flag
              properties:
                flag:
                  type: string
                  enum: [CPX, APT]
                  description: "CPX: 단지 삭제, APT: 아파트 삭제"
                  example: "CPX"
                apartmentIndex:
                  type: integer
                  description: |
                    "아파트 삭제 시 인덱스를 입력하세요 (0부터 시작).  
                    예: 1"
                  example: 1
      responses:
        "200":
          description: "삭제 성공"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "단지/아파트 삭제 성공"
        "500":
          description: "서버 오류"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "서버 오류 메시지"
...